feat: add types
Takes the types from [@types/protocol-buffers-schema](https://www.npmjs.com/package/@types/protocol-buffers-schema)
and adds them to this module.

It swaps node `Buffer`s for `Uint8Array`s in the interfaces because
this module doesn't use any `Buffer` features that aren't available
on `Uint8Array`s which makes it a bit kinder to the browser.